moved functions used only in tests
